adapter: move `sequence_peek` to a separate file
